work in progress, adding lock indication and lock/unlock set
